{{ cookiecutter.project_name }}/composition_repository/api/requirements.txt
djangorestframework>=3.15.0,<3.16.0
drf-spectacular[sidecar]>=0.28.0,<0.29.0
